+++ src/addons/meta_human_dna/__init__.py
@@ -17,7 +17,7 @@ logger = logging.getLogger(__name__)
 bl_info = {
     "name": "Meta-Human DNA",
     "author": "Poly Hammer",
-    "version": (0, 0, 15),
+    "version": (0, 1, 0),
     "blender": (4, 2, 0),
     "location": "File > Import > Metahuman DNA",
     "description": "Imports a Metahuman head from a DNA file, lets you customize it, then send it back to unreal.",
@@ -30,7 +30,8 @@ classes = [
     operators.ImportMetahumanDna,
     operators.DNA_FH_import_dna,
     operators.ConvertSelectedToDna,
-    operators.ImportMetahumanFaceAnimation,
+    operators.ImportAnimation,
+    operators.BakeAnimation,
     operators.ImportShapeKeys,
     operators.TestSentry,
     operators.OpenBuildToolDocumentation,