+++ src/addons/meta_human_dna/release_notes.md
@@ -1,5 +1,11 @@
+## Minor Changes
+* Added Animation Import onto Face Board from Unreal Face Board Level Sequence FBX [#45](https://github.com/poly-hammer/meta-human-dna-addon/issues/45)
+* Added One-click baking of face board animation down to raw controls [#63](https://github.com/poly-hammer/meta-human-dna-addon/issues/63)
+
+
 ## Patch Changes
-* Fixed matrix inverse error [#58](https://github.com/poly-hammer/meta-human-dna-addon/issues/58)
+* Fixed texture linking error on conversion
+* Fixed extra controls when re-importing face board
 
 ## Tests Passing On
 * Blender `4.2` (installed from blender.org)